daemon: consider strategy on unlock
When deciding whether to take a lock on reboot, locksmith first
considers the configured strategy. If strategy is set to BestEffort it
will then proceed to check whether the local etcd instance is active and
use this to decide whether a lock should be taken.

When deciding whether to unlock _after_ the reboot, on the other hand,
locksmith was first checking the local etcd's activeness without
considering the strategy. This leads to unexpected behaviour when
locksmith is configured to talk to a remote etcd endpoint (i.e.
LOCKSMITH_ENDPOINTS is set) and the local etcd instance is not running.

Note that this only really fixes the case of strategy being set to
always lock; the "best effort" strategy is still fundamentally broken
and we should deprecate it.